17.3:2

+++ main-service/src/main/java/ru.practicum/service/rating/RatingServiceImpl.java
@@ -15,6 +15,7 @@ import ru.practicum.dto.user.UserDto;
 import ru.practicum.entity.Event;
 import ru.practicum.entity.Rating;
 import ru.practicum.entity.User;
+import ru.practicum.enums.LikeParam;
 import ru.practicum.exception.*;
 import ru.practicum.mapper.EventMapper;
 import ru.practicum.mapper.RatingMapper;
@@ -50,10 +51,14 @@ public class RatingServiceImpl implements RatingService {
     @Override
     public OutRatingDto create(Long userId, Long eventId, IncRatingDto incRatingDto) {
         log.info("FEATURE SERVICE LOG: user id " + userId + " rate event id " + eventId);
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new NotFoundException("user id " + userId + " not found"));
-        Event event = eventRepository.findById(eventId)
-                .orElseThrow(() ->  new NotFoundException("event id " + eventId + " not found"));
+        if (!userRepository.existsById(userId)) {
+            throw new NotFoundException("user id " + userId + " not found");
+        }
+        User user = userRepository.findById(userId).get();
+        if (!eventRepository.existsById(eventId)) {
+            throw new NotFoundException("event id " + eventId + " not found");
+        }
+        Event event = eventRepository.findById(eventId).get();
         if (user.getId().equals(event.getInitiator().getId())) {
             throw new WrongRequestException("initiator can't rate event");
         }
@@ -73,10 +78,14 @@ public class RatingServiceImpl implements RatingService {
     @Override
     public OutRatingDto update(Long userId, Long ratingId, IncRatingDto incRatingDto) {
         log.info("FEATURE SERVICE LOG: private updating info about rating id " + ratingId);
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new NotFoundException("User id " + userId + " do not exist"));
-        Rating actual = ratingRepository.findById(ratingId)
-                .orElseThrow(() -> new NotFoundException("Rating id " + ratingId + " do not exist"));
+        if (!userRepository.existsById(userId)) {
+            throw new NotFoundException("user id " + userId + " not found");
+        }
+        User user = userRepository.findById(userId).get();
+        if (!validateById(ratingId)) {
+            throw new NotFoundException("rating id " + userId + " not found");
+        }
+        Rating actual = ratingRepository.findById(ratingId).get();
         if (!user.getId().equals(actual.getUser().getId())) {
             throw new IntegrityConflictException("another user can't update rating");
         }
@@ -108,8 +117,10 @@ public class RatingServiceImpl implements RatingService {
     @Override
     public OutRatingDto updateByAdmin(Long ratingId, IncRatingDto incRatingDto) {
         log.info("FEATURE SERVICE LOG: admin updating info about rating id " + ratingId);
-        Rating actual = ratingRepository.findById(ratingId)
-                .orElseThrow(() -> new NotFoundException("rating id " + ratingId + " not found"));
+        if (!validateById(ratingId)) {
+            throw new NotFoundException("rating id " + ratingId + " not found");
+        }
+        Rating actual = ratingRepository.findById(ratingId).get();
         if (Optional.ofNullable(incRatingDto.getIsLiked()).isPresent()) {
             actual.setIsLiked(incRatingDto.getIsLiked());
         }
@@ -135,7 +146,7 @@ public class RatingServiceImpl implements RatingService {
     @Override
     public void delete(Long ratingId) {
         log.info("FEATURE SERVICE LOG: delete rating id " + ratingId);
-        if (ratingRepository.existsById(ratingId)) {
+        if (validateById(ratingId)) {
             ratingRepository.deleteById(ratingId);
         }
         log.info("FEATURE SERVICE LOG: rating removed");
@@ -144,8 +155,9 @@ public class RatingServiceImpl implements RatingService {
     @Override
     public void deleteByUserId(Long userId) {
         log.info("FEATURE SERVICE LOG: removing all user id " + userId + " ratings");
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new NotFoundException("user id " + userId + " not found"));
+        if (!userRepository.existsById(userId)) {
+            throw new NotFoundException("user id " + userId + " not found");
+        }
         ratingRepository.removeAllByUserId(userId);
         log.info("FEATURE SERVICE LOG: all users rating was removed");
     }
@@ -154,8 +166,10 @@ public class RatingServiceImpl implements RatingService {
     @Transactional(readOnly = true)
     public OutRatingDto getById(Long ratingId) {
         log.info("FEATURE SERVICE LOG: getting rating id {}", ratingId);
-        Rating rating = ratingRepository.findById(ratingId)
-                .orElseThrow(() -> new NotFoundException("rating id " + ratingId + " not found"));
+        if (!validateById(ratingId)) {
+            throw new NotFoundException("rating id " + ratingId + " not found");
+        }
+        Rating rating = ratingRepository.findById(ratingId).get();
         UserDto userDto = userMapper.toUserDto(rating.getUser());
         EventShortDto eventShortDto = eventMapper
                 .toEventShortDto(rating.getEvent(), calculateEventShortRating(rating.getEvent()));
@@ -165,18 +179,21 @@ public class RatingServiceImpl implements RatingService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<OutRatingDto> getAllUsersRatings(Long userId, Boolean likesOrDislikesOnly,
+    public List<OutRatingDto> getAllUsersRatings(Long userId, LikeParam likeParam,
                                                  Integer from, Integer size) {
         log.info("FEATURE SERVICE LOG: get all user");
+        if (!userRepository.existsById(userId)) {
+            throw new NotFoundException("user id " + userId + " not found");
+        }
         CriteriaBuilder builder = entityManager.getCriteriaBuilder();
         CriteriaQuery<Rating> query = builder.createQuery(Rating.class);
 
         Root<Rating> root = query.from(Rating.class);
         Predicate criteria = builder.conjunction();
 
-        if (likesOrDislikesOnly != null) {
+        if (likeParam != null) {
             Predicate isLiked;
-            if (likesOrDislikesOnly) {
+            if (likeParam.equals(LikeParam.LIKES)) {
                 isLiked = builder.isTrue(root.get("isLiked"));
             } else {
                 isLiked = builder.isFalse(root.get("isLiked"));
@@ -184,18 +201,27 @@ public class RatingServiceImpl implements RatingService {
             criteria = builder.and(criteria, isLiked);
         }
 
+        Predicate actualUser = builder.equal(root.get("user").get("id"), userId);
+        criteria = builder.and(criteria, actualUser);
+
         query.select(root).where(criteria);
-        List<Rating> userRatings = entityManager.createQuery(query)
-                .setFirstResult(from)
-                .setMaxResults(size)
-                .getResultList();
-        log.info("FEATURE SERVICE LOG: user rating list formed; size {}", userRatings.size());
-        return ratingMapper.toOutRatingDtoList(userRatings);
+        try {
+            List<Rating> userRatings = entityManager.createQuery(query)
+                    .setFirstResult(from)
+                    .setMaxResults(size)
+                    .getResultList();
+            log.info("FEATURE SERVICE LOG: user rating list formed; size {}", userRatings.size());
+            return ratingMapper.toOutRatingDtoList(userRatings);
+        } catch (Exception e) {
+            throw new IntegrityConflictException("query exception");
+        }
     }
 
     @Override
     @Transactional(readOnly = true)
-    public List<OutRatingDto> getAll(Boolean likesOrDislikesOnly,
+    public List<OutRatingDto> getAll(LikeParam likeParam,
+                                     List<Long> userIds,
+                                     List<Long> eventIds,
                                      Integer from,
                                      Integer size) {
         log.info("FEATURE SERVICE LOG: get all rating list with params");
@@ -205,23 +231,38 @@ public class RatingServiceImpl implements RatingService {
         Root<Rating> root = query.from(Rating.class);
         Predicate criteria = builder.conjunction();
 
-        if (likesOrDislikesOnly != null) {
+        if (likeParam != null) {
             Predicate isLiked;
-            if (likesOrDislikesOnly) {
+            if (likeParam.equals(LikeParam.LIKES)) {
+                log.info("like params: " + likeParam);
                 isLiked = builder.isTrue(root.get("isLiked"));
             } else {
+                log.info("like params: " + likeParam);
                 isLiked = builder.isFalse(root.get("isLiked"));
             }
             criteria = builder.and(criteria, isLiked);
         }
+        if (Optional.ofNullable(userIds).isPresent() && userIds.size() > 0) {
+            Predicate containUsers = root.get("user").get("id").in(userIds);
+            criteria = builder.and(criteria, containUsers);
+        }
+
+        if (Optional.ofNullable(eventIds).isPresent() && eventIds.size() > 0) {
+            Predicate containEvents = root.get("event").get("id").in(eventIds);
+            criteria = builder.and(criteria, containEvents);
+        }
 
         query.select(root).where(criteria);
-        List<Rating> ratings = entityManager.createQuery(query)
-                .setFirstResult(from)
-                .setMaxResults(size)
-                .getResultList();
-        log.info("FEATURE SERVICE LOG: rating list formed");
-        return ratingMapper.toOutRatingDtoList(ratings);
+        try {
+            List<Rating> ratings = entityManager.createQuery(query)
+                    .setFirstResult(from)
+                    .setMaxResults(size)
+                    .getResultList();
+            log.info("FEATURE SERVICE LOG: rating list formed");
+            return ratingMapper.toOutRatingDtoList(ratings);
+        } catch (Exception e) {
+            throw new IntegrityConflictException("query exception");
+        }
     }
 
     @Override
@@ -262,53 +303,55 @@ public class RatingServiceImpl implements RatingService {
         return eventShortDtos;
     }
 
+    @Override
     public EventFullRatingDto calculateEventRating(Event event) {
-            List<Rating> eventRatings = event.getRatingList();
-            int likesCount = 0;
-            int dislikesCount = 0;
-            float contentSum = 0.0f;
-            float locationSum = 0.0f;
-            float organizationSum = 0.0f;
-            int contentCount = 0;
-            int locationCount = 0;
-            int organizationCount = 0;
-            for (Rating rating : eventRatings) {
-                if (Optional.ofNullable(rating.getContentRate()).isPresent()) {
-                    contentSum += rating.getContentRate();
-                    contentCount++;
-                }
-                if (Optional.ofNullable(rating.getOrganizationRate()).isPresent()) {
-                    organizationSum += rating.getOrganizationRate();
-                    organizationCount++;
-                }
-                if (Optional.ofNullable(rating.getLocationRate()).isPresent()) {
-                    locationSum += rating.getLocationRate();
-                    locationCount++;
-                }
-                if (Optional.ofNullable(rating.getIsLiked()).isPresent()) {
-                    if (rating.getIsLiked().equals(Boolean.TRUE)) {
-                        likesCount++;
-                    } else {
-                        dislikesCount++;
-                    }
+        List<Rating> eventRatings = event.getRatingList();
+        int likesCount = 0;
+        int dislikesCount = 0;
+        float contentSum = 0.0f;
+        float locationSum = 0.0f;
+        float organizationSum = 0.0f;
+        int contentCount = 0;
+        int locationCount = 0;
+        int organizationCount = 0;
+        for (Rating rating : eventRatings) {
+            if (Objects.nonNull(rating.getContentRate())) {
+                contentSum += rating.getContentRate();
+                contentCount++;
+            }
+            if (Objects.nonNull(rating.getOrganizationRate())) {
+                organizationSum += rating.getOrganizationRate();
+                organizationCount++;
+            }
+            if (Objects.nonNull(rating.getLocationRate())) {
+                locationSum += rating.getLocationRate();
+                locationCount++;
+            }
+            if (Objects.nonNull(rating.getIsLiked())) {
+                if (rating.getIsLiked().equals(Boolean.TRUE)) {
+                    likesCount++;
+                } else {
+                    dislikesCount++;
                 }
             }
-            int percentRate = (likesCount == 0 && dislikesCount == 0) ? 0 : (likesCount * 100) / (dislikesCount + likesCount);
-            EventFullRatingDto eventFullRatingDto = EventFullRatingDto.builder()
-                    .contentRate(contentSum / contentCount)
-                    .locationRate(locationSum / locationCount)
-                    .organizationRate(organizationSum / organizationCount)
-                    .percentRating(percentRate)
-                    .build();
+        }
+        int percentRate = (likesCount == 0 && dislikesCount == 0) ? 0 : (likesCount * 100) / (dislikesCount + likesCount);
+        EventFullRatingDto eventFullRatingDto = EventFullRatingDto.builder()
+                .contentRate(contentSum / contentCount)
+                .locationRate(locationSum / locationCount)
+                .organizationRate(organizationSum / organizationCount)
+                .percentRating(percentRate)
+                .build();
         return eventFullRatingDto;
     }
 
+    @Override
     public EventShortRatingDto calculateEventShortRating(Event event) {
         List<Rating> eventRatings = event.getRatingList();
         int likesCount = 0;
         int dislikesCount = 0;
         for (Rating rating : eventRatings) {
-            if (Optional.ofNullable(rating.getIsLiked()).isPresent()) {
+            if (Objects.nonNull(rating.getIsLiked())) {
                 if (rating.getIsLiked().equals(Boolean.TRUE)) {
                     likesCount++;
                 } else {
@@ -318,5 +361,9 @@ public class RatingServiceImpl implements RatingService {
         }
         int percentRate = (likesCount == 0 && dislikesCount == 0) ? 0 : (likesCount * 100) / (dislikesCount + likesCount);
         return EventShortRatingDto.builder().percentRating(percentRate).build();
+    }
+
+    private Boolean validateById(Long id) {
+        return ratingRepository.existsById(id);
     }
 }
